GCOV-11 Set SonarQube version to 7.9 GCOV-10 Add integration tests GCOV-8 Set parent POM to org.sonarsource.parent:parent GCOV-7 Move packages to com.lintyservices GCOV-5 Update metadata GCOV-4 Compile with JDK 11 GCOV-3 Update license headers GCOV-2 Set version to 1.3.0-SNAPSHOT
