+++ .gitignore
@@ -1,4 +1,4 @@
-/target/
+target
 /sonar-project.properties
 .classpath
 .project
