+++ gcov-its/pom.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>com.lintyservices.sonar.plugins.gcov</groupId>
+    <artifactId>parent</artifactId>
+    <version>1.3.0-SNAPSHOT</version>
+  </parent>
+
+  <artifactId>gcov-its</artifactId>
+  <packaging>pom</packaging>
+  <name>SonarQube Linty Gcov :: Integration Tests</name>
+
+  <modules>
+    <module>plugin</module>
+  </modules>
+
+</project>
