+++ sonar-gcov-plugin/src/test/java/com/lintyservices/sonar/plugins/gcov/GcovPluginTest.java
@@ -1,0 +1,50 @@
+/*
+ * SonarQube Linty Gcov :: Plugin
+ * Copyright (C) 2019-2020 Linty Services
+ * mailto:contact@linty-services.com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package com.lintyservices.sonar.plugins.gcov;
+
+import org.junit.Test;
+import org.sonar.api.Plugin;
+import org.sonar.api.SonarEdition;
+import org.sonar.api.SonarQubeSide;
+import org.sonar.api.SonarRuntime;
+import org.sonar.api.internal.SonarRuntimeImpl;
+import org.sonar.api.utils.Version;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class GcovPluginTest {
+
+  public static final Version LTS_VERSION = Version.create(7, 9);
+  private static final int EXTENSIONS = 2;
+
+  @Test
+  public void should_contain_the_right_number_of_extensions() {
+    Plugin.Context context = setupContext(SonarRuntimeImpl.forSonarQube(LTS_VERSION, SonarQubeSide.SERVER, SonarEdition.COMMUNITY));
+    assertThat(context.getExtensions()).hasSize(EXTENSIONS);
+  }
+
+  private Plugin.Context setupContext(SonarRuntime runtime) {
+    Plugin.Context context = new Plugin.Context(runtime);
+    new GcovPlugin().define(context);
+    return context;
+  }
+
+  //TODO : test properties
+}
