+++ sonar-gcov-plugin/src/test/resources/com/lintyservices/sonar/plugins/gcov/GcovSensorTests/bad-coverage.gcov
@@ -1,0 +1,3 @@
+aaaaaaaaaaaaaaaaaaaaaaaaa
+aaaaaaaaaaaaaaaaaaa
+123
