+++ sonar-gcov-plugin/src/test/resources/com/lintyservices/sonar/plugins/gcov/GcovSensorTests/badvalue-coverage.gcov
@@ -1,0 +1,5 @@
+        -:    0:Source:D:/Stage/2018-001-test_coverage_sonarqube/root/WorkSpaceMentor/plasma/plasma_custom/dev/src/cpu/alu.vhd
+        -:    0:Object:alu.bb
+        -:    1:-------------------------------------------------------------------------------
+        aaaa:    2:--
+        -:    3:-- The Arithmetic Logic Unit (ALU).
