+++ sonar-gcov-plugin/src/test/resources/com/lintyservices/sonar/plugins/gcov/GcovSensorTests/commons-chain-coverage.gcov
