--- src/test/resources/org/sonar/plugins/gcov/GcovSensorTests/bad-coverage.gcov
@@ -1,3 +1,0 @@
-aaaaaaaaaaaaaaaaaaaaaaaaa
-aaaaaaaaaaaaaaaaaaa
-123
